Separate constraint translation algorithm from Swift code generation
