+++ compiler/core/src/utils/constraint.bs.js
@@ -2,8 +2,12 @@
 'use strict';
 
 var $$Map                   = require("bs-platform/lib/js/map.js");
+var List                    = require("bs-platform/lib/js/list.js");
+var Block                   = require("bs-platform/lib/js/block.js");
 var Curry                   = require("bs-platform/lib/js/curry.js");
 var Caml_obj                = require("bs-platform/lib/js/caml_obj.js");
+var Pervasives              = require("bs-platform/lib/js/pervasives.js");
+var Layer$LonaCompilerCore  = require("../core/layer.bs.js");
 var Caml_builtin_exceptions = require("bs-platform/lib/js/caml_builtin_exceptions.js");
 
 function anchorToString(param) {
@@ -226,6 +230,194 @@ var ConstraintMap = /* module */[
   /* find_opt */find_opt
 ];
 
+function getConstraints(rootLayer) {
+  var constrainAxes = function (layer) {
+    var direction = Layer$LonaCompilerCore.getFlexDirection(layer);
+    var isColumn = +(direction === "column");
+    var primaryBeforeAnchor = isColumn !== 0 ? /* Top */2 : /* Leading */4;
+    var primaryAfterAnchor = isColumn !== 0 ? /* Bottom */3 : /* Trailing */5;
+    var secondaryBeforeAnchor = isColumn !== 0 ? /* Leading */4 : /* Top */2;
+    var secondaryAfterAnchor = isColumn !== 0 ? /* Trailing */5 : /* Bottom */3;
+    var primaryDimensionAnchor = isColumn !== 0 ? /* Height */1 : /* Width */0;
+    var secondaryDimensionAnchor = isColumn !== 0 ? /* Width */0 : /* Height */1;
+    var height = Layer$LonaCompilerCore.getNumberParameterOpt("height", layer);
+    var width = Layer$LonaCompilerCore.getNumberParameterOpt("width", layer);
+    var sizingRules = Layer$LonaCompilerCore.getSizingRules(Layer$LonaCompilerCore.findParent(rootLayer, layer), layer);
+    var primarySizingRule = isColumn !== 0 ? sizingRules[/* height */1] : sizingRules[/* width */0];
+    var secondarySizingRule = isColumn !== 0 ? sizingRules[/* width */0] : sizingRules[/* height */1];
+    var flexChildren = List.filter((function (child) {
+              return +(Layer$LonaCompilerCore.getNumberParameter("flex", child) === 1.0);
+            }))(layer[/* children */3]);
+    var addConstraints = function (index, child) {
+      var childSizingRules = Layer$LonaCompilerCore.getSizingRules(/* Some */[layer], child);
+      var childSecondarySizingRule = isColumn !== 0 ? childSizingRules[/* width */0] : childSizingRules[/* height */1];
+      var firstViewConstraints = index !== 0 ? /* [] */0 : /* :: */[
+          /* Relation */Block.__(1, [
+              child,
+              primaryBeforeAnchor,
+              /* Eq */0,
+              layer,
+              primaryBeforeAnchor,
+              /* Required */0,
+              /* PrimaryBefore */0
+            ]),
+          /* [] */0
+        ];
+      var lastViewConstraints;
+      if (index === (List.length(layer[/* children */3]) - 1 | 0)) {
+        var match = List.length(flexChildren);
+        var needsPrimaryAfterConstraint = typeof primarySizingRule === "number" ? (
+            primarySizingRule !== 0 || match ? /* true */1 : /* false */0
+          ) : (
+            match ? /* true */1 : /* false */0
+          );
+        lastViewConstraints = needsPrimaryAfterConstraint !== 0 ? /* :: */[
+            /* Relation */Block.__(1, [
+                child,
+                primaryAfterAnchor,
+                /* Eq */0,
+                layer,
+                primaryAfterAnchor,
+                /* Required */0,
+                /* PrimaryAfter */2
+              ]),
+            /* [] */0
+          ] : /* [] */0;
+      } else {
+        lastViewConstraints = /* [] */0;
+      }
+      var middleViewConstraints;
+      if (index !== 0) {
+        var previousLayer = List.nth(layer[/* children */3], index - 1 | 0);
+        middleViewConstraints = /* :: */[
+          /* Relation */Block.__(1, [
+              child,
+              primaryBeforeAnchor,
+              /* Eq */0,
+              previousLayer,
+              primaryAfterAnchor,
+              /* Required */0,
+              /* PrimaryBetween */1
+            ]),
+          /* [] */0
+        ];
+      } else {
+        middleViewConstraints = /* [] */0;
+      }
+      var secondaryBeforeConstraint = /* Relation */Block.__(1, [
+          child,
+          secondaryBeforeAnchor,
+          /* Eq */0,
+          layer,
+          secondaryBeforeAnchor,
+          /* Required */0,
+          /* SecondaryBefore */3
+        ]);
+      var secondaryAfterConstraint;
+      var exit = 0;
+      if (typeof secondarySizingRule === "number" && !(secondarySizingRule !== 0 || !(typeof childSecondarySizingRule === "number" && childSecondarySizingRule !== 0))) {
+        secondaryAfterConstraint = /* :: */[
+          /* Relation */Block.__(1, [
+              child,
+              secondaryAfterAnchor,
+              /* Leq */2,
+              layer,
+              secondaryAfterAnchor,
+              /* Required */0,
+              /* SecondaryAfter */4
+            ]),
+          /* [] */0
+        ];
+      } else {
+        exit = 1;
+      }
+      if (exit === 1) {
+        secondaryAfterConstraint = typeof childSecondarySizingRule === "number" ? /* :: */[
+            /* Relation */Block.__(1, [
+                child,
+                secondaryAfterAnchor,
+                /* Eq */0,
+                layer,
+                secondaryAfterAnchor,
+                /* Required */0,
+                /* SecondaryAfter */4
+              ]),
+            /* [] */0
+          ] : /* [] */0;
+      }
+      var fitContentSecondaryConstraint = typeof secondarySizingRule === "number" && secondarySizingRule !== 0 ? /* :: */[
+          /* Relation */Block.__(1, [
+              child,
+              secondaryDimensionAnchor,
+              /* Leq */2,
+              layer,
+              secondaryDimensionAnchor,
+              /* Low */1,
+              /* FitContentSecondary */5
+            ]),
+          /* [] */0
+        ] : /* [] */0;
+      return Pervasives.$at(firstViewConstraints, Pervasives.$at(lastViewConstraints, Pervasives.$at(middleViewConstraints, Pervasives.$at(/* :: */[
+                              secondaryBeforeConstraint,
+                              /* [] */0
+                            ], Pervasives.$at(secondaryAfterConstraint, fitContentSecondaryConstraint)))));
+    };
+    var flexChildrenConstraints;
+    if (flexChildren) {
+      var rest = flexChildren[1];
+      var first = flexChildren[0];
+      if (List.length(rest) > 0) {
+        flexChildrenConstraints = List.map((function (param) {
+                var anchor = primaryDimensionAnchor;
+                var layer = param;
+                return /* Relation */Block.__(1, [
+                          first,
+                          anchor,
+                          /* Eq */0,
+                          layer,
+                          anchor,
+                          /* Required */0,
+                          /* FlexSibling */6
+                        ]);
+              }), rest);
+      } else {
+        flexChildrenConstraints = /* [] */0;
+      }
+    } else {
+      flexChildrenConstraints = /* [] */0;
+    }
+    var heightConstraint = height ? /* :: */[
+        /* Dimension */Block.__(0, [
+            layer,
+            /* Height */1,
+            /* Required */0,
+            isColumn !== 0 ? /* PrimaryDimension */7 : /* SecondaryDimension */8
+          ]),
+        /* [] */0
+      ] : /* [] */0;
+    var widthConstraint = width ? /* :: */[
+        /* Dimension */Block.__(0, [
+            layer,
+            /* Width */0,
+            /* Required */0,
+            isColumn !== 0 ? /* SecondaryDimension */8 : /* PrimaryDimension */7
+          ]),
+        /* [] */0
+      ] : /* [] */0;
+    return List.concat(Pervasives.$at(/* :: */[
+                    heightConstraint,
+                    /* :: */[
+                      widthConstraint,
+                      /* [] */0
+                    ]
+                  ], Pervasives.$at(/* :: */[
+                        flexChildrenConstraints,
+                        /* [] */0
+                      ], List.mapi(addConstraints, layer[/* children */3]))));
+  };
+  return List.concat(Layer$LonaCompilerCore.flatmap(constrainAxes, rootLayer));
+}
+
 exports.anchorToString   = anchorToString;
 exports.anchorFromString = anchorFromString;
 exports.cmpFromString    = cmpFromString;
@@ -233,4 +425,5 @@ exports.cmpToString      = cmpToString;
 exports.getPriority      = getPriority;
 exports.getRole          = getRole;
 exports.ConstraintMap    = ConstraintMap;
+exports.getConstraints   = getConstraints;
 /* include Not a pure module */
